Small fixes, did not fix the bug
There currently exists a bug which is replicated in the following way:
Person A spawns on the floor the elevator is on.
Person B spawns on another floor.
B is faster than A.
The elevator will start heading up to B as they are ready first.
A will still have their permit and will head straight into a closed door.
A is displeased.